Using refs to focus the next account selector
This was tricky to get working!
Had to use a Map (which guarantees insertion order) to store the refs. And my desired impl of focussing the next txn was just too complicated to get right, so I'm just focussing on the first one.
The index for the compound ref is very important that it's consistent. I kept using inconsistent ones (the index in the rendering list) and getting weird results.